requirements.txt
streamlit
torch
torchvision
scikit-learn
Pillow
streamlit-qr-code-scanner